Ask before install the Oh my ZSH part

+++ install.sh
@@ -10,8 +10,10 @@ echo "** Installing apt packages"
 sudo apt-get update
 sudo apt-get install -y --no-install-recommends zsh fzf vim jq unzip
 
+# ZSH
 USER=`whoami`
 sudo -n chsh $USER -s $(which zsh)
+ln -sf ~/timer-dotfiles/config/.zshrc ~/.zshrc
 
 #----------------------------------------------------------------------------
 # GH CLI
@@ -25,25 +27,31 @@ curl -s https://api.github.com/repos/cli/cli/releases/latest \
 sudo -n dpkg -i ./gh_*.deb
 rm ./gh_*.deb
 
-# ZSH
-ln -sf ~/timer-dotfiles/config/.zshrc ~/.zshrc
-
 #----------------------------------------------------------------------------
 # Oh My ZSH
 #----------------------------------------------------------------------------
-echo
-echo "** Installing Oh My Zsh"
-rm -rf ~/.oh-my-zsh
-touch ~/.z  # So it doesn't complain on very first usage
-CHSH=no RUNZSH=no KEEP_ZSHRC=yes sh -c "$(curl -fsSL https://raw.github.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
 
-# Oh my ZSH theme
-git clone https://github.com/romkatv/powerlevel10k.git ~/.oh-my-zsh/custom/themes/powerlevel10k --depth 1
-ln -sf ~/timer-dotfiles/config/.p10k.zsh ~/.p10k.zsh
+# Ask the user if they want to install Oh My Zsh
+read -r -p "Do you want to install Oh My Zsh? (y/n): " install_oh_my_zsh
 
-# Oh my ZSH plugin
-git clone https://github.com/zsh-users/zsh-autosuggestions.git ~/.oh-my-zsh/custom/plugins/zsh-autosuggestions --depth 1
-git clone https://github.com/zsh-users/zsh-syntax-highlighting.git ~/.oh-my-zsh/custom/plugins/zsh-syntax-highlighting --depth 1
+if [ "$install_oh_my_zsh" == "y" ]; then
+  echo "** Installing Oh My Zsh"
+
+  # Oh my ZSH - tool
+  rm -rf ~/.oh-my-zsh
+  touch ~/.z  # So it doesn't complain on very first usage
+  CHSH=no RUNZSH=no KEEP_ZSHRC=yes sh -c "$(curl -fsSL https://raw.github.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
+
+  # Oh my ZSH - theme
+  git clone https://github.com/romkatv/powerlevel10k.git ~/.oh-my-zsh/custom/themes/powerlevel10k --depth 1
+  ln -sf ~/timer-dotfiles/config/.p10k.zsh ~/.p10k.zsh
+
+  # Oh my ZSH - plugin
+  git clone https://github.com/zsh-users/zsh-autosuggestions.git ~/.oh-my-zsh/custom/plugins/zsh-autosuggestions --depth 1
+  git clone https://github.com/zsh-users/zsh-syntax-highlighting.git ~/.oh-my-zsh/custom/plugins/zsh-syntax-highlighting --depth 1
+else
+  echo "Skipping Oh My Zsh installation."
+fi
 
 #----------------------------------------------------------------------------
 # pyenv
